pyproject.toml
[project]
name = "origin-data-infra"
version = "0.1.0"
description = "Origin Transport Data Infrastructure - Supabase + Google Sheets sync"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Database & Sync
    "supabase>=2.25.0",
    "gspread>=6.2.1",
    "google-auth>=2.37.0",
    # GCP
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=3.6.0",
    "pandas>=2.3.2",
    "db-dtypes>=1.3.0",  # BigQuery -> pandas type support
    # AI
    "anthropic>=0.75.0",
    # Banking
    "plaid-python>=37.1.0",
    # Accounting
    "python-quickbooks>=0.9.12",
    "intuit-oauth>=1.2.6",
    # Core
    "python-dotenv>=1.0.1",
    "pydantic>=2.12.5",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
    "mypy>=1.14.1",
]

[project.scripts]
origin = "scripts.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
